Campus Courier Website
This website is designed for people for ordering food and getting
delivery at their desired place within UNC Charlotte
campus. Firstly, the customer surfs for food courts we have at our
Uniersity and places an order, enters the destination
and makes payment using PayPal. Then we have students registered to the
website as Couriers who are willing to deliver
the orders to the customers, in return they will earn some money. The
courier will login to the system and will be able
to see the list of orders and will select the order he wish to deliver.
We maintained few statuses of the orders so
that various couriers can differentiate which order to deliver.
This was a group project with team size 3.
I have done ordering food module and implemented PayPal functionality
using PayPal API. Future enhancement for this project
are Real time tracking of the order and Delivering documents or rides
among the University Campus.